+++ src/assembler/builtin-calls/compile-export.ts
@@ -1,5 +1,5 @@
-import { CompileExprOpts, compileExpression } from "../assembler.js";
-import { Call } from "../syntax-objects/call.js";
+import { CompileExprOpts, compileExpression } from "../../assembler.js";
+import { Call } from "../../syntax-objects/call.js";
 
 export const compileExport = (opts: CompileExprOpts<Call>) => {
   const expr = opts.expr.exprArgAt(0);
