+++ src/assembler/builtin-calls/compile-member-access.ts
@@ -1,9 +1,9 @@
-import { CompileExprOpts, compileExpression, mapBinaryenType } from "../assembler.js";
-import { Call } from "../syntax-objects/call.js";
-import { ObjectType, IntersectionType } from "../syntax-objects/types.js";
-import { getExprType } from "../semantics/resolution/get-expr-type.js";
+import { CompileExprOpts, compileExpression, mapBinaryenType } from "../../assembler.js";
+import { Call } from "../../syntax-objects/call.js";
+import { ObjectType, IntersectionType } from "../../syntax-objects/types.js";
+import { getExprType } from "../../semantics/resolution/get-expr-type.js";
 import { OBJECT_FIELDS_OFFSET } from "./object-layout.js";
-import * as gc from "../lib/binaryen-gc/index.js";
+import * as gc from "../../lib/binaryen-gc/index.js";
 
 export const compileMemberAccess = (opts: CompileExprOpts<Call>) => {
   const { expr, mod } = opts;
