+++ src/assembler/builtin-calls/object-layout.ts
